requirements.txt
azure-devops==6.0.0b4
Flask==2.0.1
bs4==0.0.1
Markdown==3.3.4
requests==2.26.0
Jinja2==3.0.1
python-dotenv==0.19.0